Consider all selection ranges when determining if mark is set
